chore: move form-los component
