+++ src/components/form-los/configured-form-los.jsx
@@ -1,7 +1,7 @@
 import { reduxForm } from 'redux-form';
 
-import { FormLOS } from './form-los.component';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import FormLOS from './form-los.component';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'themes',
