+++ src/components/form-los/form-los.scss
